requirements.txt
streamlit>=1.32.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
openai>=1.12.0
scikit-learn>=1.4.0
nltk>=3.8.1
requests>=2.31.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
reportlab>=4.0.0
python-docx>=1.0.0
